+++ Controller/src/main.rs
@@ -26,6 +26,7 @@ use log::{info, warn};
 use macaddr::MacAddr;
 use rbfrt::error::RBFRTError;
 use rbfrt::util::port_manager::{AutoNegotiation, FEC, Loopback, Port, Speed};
+use rbfrt::util::port_manager::FEC::BF_FEC_TYP_REED_SOLOMON;
 use rbfrt::util::PortManager;
 use tokio::sync::Mutex;
 
@@ -35,6 +36,7 @@ mod error;
 
 use core::FrameSizeMonitor;
 use crate::core::{Arp, Config, FrameTypeMonitor, RateMonitor, TrafficGen};
+use crate::core::traffic_gen_core::const_definitions::{PORT_CFG_TF2};
 use crate::core::traffic_gen_core::event::TrafficGenEvent;
 
 #[derive(Debug, Copy, Clone)]
@@ -62,10 +64,14 @@ pub struct AppState {
     pub(crate) experiment: Mutex<Experiment>,
     pub(crate) sample_mode: bool,
     pub(crate) config: Mutex<Config>,
-    pub(crate) arp_handler: Arp
+    pub(crate) arp_handler: Arp,
+    pub(crate) tofino2: bool,
+    pub(crate) loopback_mode: bool
 }
 
-async fn configure_ports(switch: &mut SwitchConnection, pm: &PortManager, config: &Config, recirculation_ports: &Vec<u32>, port_mapping: &mut HashMap<u32, PortMapping>) -> Result<(), RBFRTError> {
+async fn configure_ports(switch: &mut SwitchConnection, pm: &PortManager, config: &Config,
+                         recirculation_ports: &Vec<u32>, port_mapping: &mut HashMap<u32, PortMapping>,
+                         is_tofino2: bool, loopback_mode: bool) -> Result<(), RBFRTError> {
     // Delete previously configured ports
     switch.clear_table("$PORT").await?;
 
@@ -74,10 +80,14 @@ async fn configure_ports(switch: &mut SwitchConnection, pm: &PortManager, config
 
     // TG_PORTS
     for tg in &config.tg_ports {
-        let pm_req = Port::new(tg.port , 0)
-            .speed(Speed::BF_SPEED_100G)
-            .fec(FEC::BF_FEC_TYP_NONE)
+        let mut pm_req = Port::new(tg.port , 0)
+            .speed(if is_tofino2 {Speed::BF_SPEED_400G} else {Speed::BF_SPEED_100G})
+            .fec(if is_tofino2 {BF_FEC_TYP_REED_SOLOMON} else {FEC::BF_FEC_TYP_NONE})
             .auto_negotiation(AutoNegotiation::PM_AN_DEFAULT);
+
+        if loopback_mode { // loopback mode is used for testing if no cables are available
+            pm_req = pm_req.loopback(Loopback::BF_LPBK_MAC_NEAR);
+        }
 
         // we validated the mac address before
         tg_ports.push((tg.port, MacAddr::from_str(&tg.mac).unwrap()));
@@ -88,8 +98,8 @@ async fn configure_ports(switch: &mut SwitchConnection, pm: &PortManager, config
     // Recirculation ports
     for port in recirculation_ports {
         let pm_req = Port::new(*port , 0)
-            .speed(Speed::BF_SPEED_100G)
-            .fec(FEC::BF_FEC_TYP_NONE)
+            .speed(if is_tofino2 {Speed::BF_SPEED_400G} else {Speed::BF_SPEED_100G})
+            .fec(if is_tofino2 {BF_FEC_TYP_REED_SOLOMON} else {FEC::BF_FEC_TYP_NONE})
             .auto_negotiation(AutoNegotiation::PM_AN_DEFAULT)
             .loopback(Loopback::BF_LPBK_MAC_NEAR);
 
@@ -131,6 +141,7 @@ async fn run() -> Result<(), Box<dyn std::error::Error>> {
     let sample_mode = env::var("SAMPLE").unwrap_or("0".to_owned()).parse().unwrap_or(0);
     let sample_mode = sample_mode == 1;
     let p4_name = env::var("P4_NAME").unwrap_or("traffic_gen".to_owned());
+    let loopback_mode = env::var("LOOPBACK").unwrap_or("0".to_owned()).parse().unwrap_or(false);
 
     info!("Start controller...");
 
@@ -140,6 +151,22 @@ async fn run() -> Result<(), Box<dyn std::error::Error>> {
         .p4_name(&p4_name)
         .connect()
         .await?;
+
+    // check if its tofino 1 or tofino 2
+    // this could be done more intelligent
+    // we simply check if a table in tf2 scope exists
+    let is_tofino2 = switch.has_table(PORT_CFG_TF2);
+
+    if is_tofino2 {
+        info!("ASIC: Tofino2");
+    }
+    else {
+        info!("ASIC: Tofino1");
+    }
+
+    if loopback_mode {
+        info!("Loopback mode activated.");
+    }
 
     // Front panel ports that can be used for traffic generation.
     // At default, the first 10 ports are used for traffic generation.
@@ -182,7 +209,7 @@ async fn run() -> Result<(), Box<dyn std::error::Error>> {
 
     let pm = PortManager::new(&switch).await;
 
-    configure_ports(&mut switch, &pm, &config, &recirculation_ports, &mut port_mapping).await?;
+    configure_ports(&mut switch, &pm, &config, &recirculation_ports, &mut port_mapping, is_tofino2, loopback_mode).await?;
 
     // configures frame size count tables
     let frame_size_monitor = FrameSizeMonitor::new(port_mapping.clone());
@@ -196,7 +223,7 @@ async fn run() -> Result<(), Box<dyn std::error::Error>> {
     rate_monitor.init_iat_meter(&switch, sample_mode).await?;
     rate_monitor.on_reset(&switch).await?;
 
-    let mut traffic_generator = TrafficGen::new();
+    let mut traffic_generator = TrafficGen::new(is_tofino2);
     traffic_generator.stop(&switch).await?;
 
     let index_mapping = traffic_generator.init_monitoring_packet(&switch, &port_mapping).await?;
@@ -215,7 +242,9 @@ async fn run() -> Result<(), Box<dyn std::error::Error>> {
         sample_mode,
         experiment: Mutex::new(Experiment { start: std::time::SystemTime::now(), running: false }),
         config: Mutex::new(config),
-        arp_handler
+        arp_handler,
+        tofino2: is_tofino2,
+        loopback_mode
     });
 
     state.frame_size_monitor.lock().await.configure(&state.switch).await?;